llamafarms/requirements.txt
# Core AI and ML requirements
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
tensorflow>=2.10.0
mlx>=0.0.1  # Apple MLX for accelerated ML on Apple Silicon
mlx-image>=0.0.1  # Image models optimized for MLX

# Foundation models and APIs
openai>=1.0.0  # OpenAI API client
langchain>=0.0.1  # Framework for LLM applications
transformers>=4.30.0  # Hugging Face Transformers
datasets>=2.10.0  # Hugging Face Datasets
sentence-transformers>=2.2.0  # Sentence embeddings

# Computer vision
pillow>=9.0.0
opencv-python>=4.5.0
ultralytics>=8.0.0  # YOLO v8
segment-anything>=1.0  # SAM integration

# Satellite imaging and geospatial
sentinelhub>=3.4.0
rasterio>=1.2.0
geopandas>=0.9.0
shapely>=1.7.0
pyproj>=3.1.0
geopy>=2.2.0

# API and web backend
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.5
aiofiles>=0.7.0
sqlalchemy>=2.0.0
httpx>=0.24.0  # HTTP client

# Testing and quality assurance
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.20.0
pytest-mock>=3.10.0
black>=23.0.0  # Code formatter
flake8>=6.0.0  # Linter
mypy>=1.0.0  # Type checker

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
mkdocs>=1.4.0
mkdocs-material>=9.0.0

# Data visualization
plotly>=5.13.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0  # Progress bars
tenacity>=8.2.0  # Retry logic
loguru>=0.6.0  # Logging
click>=8.1.0  # Command-line interfaces
python-dotenv>=1.0.0  # Environment variables
requests>=2.28.0  # HTTP requests 